changed how the reply looks and added location name

+++ Project_Folder/Input_Page/functions/delete_location.php
@@ -6,17 +6,28 @@
 
 require './db_connect.php';
 
-if(isset($_POST['LocationID'])){
+if(isset($_POST['LocationID']) && isset($_POST['LocationName'])){
     $locationID = $_POST['LocationID'];
+    $locationName = $_POST['LocationName'];
     
-    if(!empty($locationID)){
+    if(!empty($locationID) && !empty($locationName)){
         $delete_query = "DELETE FROM locationdata WHERE LocationID=".$locationID;
         $delete_query_results = mysqli_query($connection, $delete_query);
         
         if($delete_query_results){
-            echo 'The location <b>'.$locationName.'</b> has been deleted from the databse';
+            echo '<div id="message-box-error" class="message-box">
+                        <a class="close fade-out">×</a>
+                        <p class="text-center">
+                            The location <strong>'.$locationName.'</strong> has been deleted.
+                        </p>
+                     </div>';
         }else{
-            echo 'Error: The location <b>'.$locationName.'</b> could not be deleted from the database';
+            echo '<div id="message-box-error" class="message-box">
+                        <a class="close fade-out">×</a>
+                        <p class="text-center">
+                            <strong>Error:</strong> The location <strong>'.$locationName.'</strong> could not be deleted.
+                        </p>
+                     </div>';
         }
     }
 }
